requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==3.0
Django==3.0.7
django-crispy-forms==1.9.1
django-fernet-fields==0.6
idna==2.9
Pillow==7.2.0
pycparser==2.20
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
sip==4.19.8
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
